scripts/requirements.txt
boto==2.49.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
distro==1.5.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-oauth2-tool==0.0.3
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
httplib2==0.17.1
idna==2.9
oauth2client==4.1.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
